Move boundaries and centers functions to utils.py
